layered_grpc/data_service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
redis==5.0.8
protobuf==4.25.1
